backend/requirements.txt
flask==3.0.3
pillow==10.3.0
opencv-contrib-python==4.6.0.66
numpy==1.26.4
pandas==2.2.2
transformers==4.40.2
datasets==2.19.1
torch==2.2.2+cu121
torchvision==0.17.2+cu121
torchaudio==2.2.2+cu121
--extra-index-url https://download.pytorch.org/whl/cu121
pdf2image==1.17.0
roboflow==1.1.27
python-dotenv==1.0.1
accelerate==0.29.3
seqeval==1.2.2
ipywidgets
groq
sentencepiece
flask-cors==4.0.0

# --- OCR dependencies required for paddleocr ---
attrdict
beautifulsoup4
cython
fire
imgaug
lmdb
lxml
openpyxl
pdf2docx
premailer
pyclipper
python-docx
rapidfuzz
scikit-image
visualdl

# Install paddleocr (core package) without dependencies:
# pip install paddleocr==2.7.0.3 --no-deps

# Install paddlepaddle-gpu matching your CUDA version (example below for CUDA 11.2/11.6/12.1):
# pip install paddlepaddle-gpu==2.6.2 -f https://www.paddlepaddle.org.cn/whl/windows/mkl/avx/stable.html
